Moved defaults into pkg/util/crd_defaults.go and renamed synopsysctl defaults-flags file to commandflags.go
